refactor: Loading component path refatoring
